Create HUD and add unit placement
Add framework for placing units

Add buttons that can be pressed

In the future they will create units, but currently they just print something

make units spawn when enter is pressed

And a button is selected. I had to remove the ui_accept inputs from the
input map, because they would toggle the button when enter was pressed.